backend/python/requirement.txt
fastapi[standard]
pandas
sdmx1
pysdmx
